Amended broken view models into its whole files
